tools/extract/gemini/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
google-genai==1.2.0
requests==2.28.2
tenacity==8.2.3
google-generativeai==0.8.4
